First working version of copy case section feature. Renames & moves
