+++ plugin-hrm-form/src/components/case/AddEditCaseItem.tsx
@@ -48,8 +48,9 @@ import {
   temporaryCaseInfoHistory,
 } from '../../states/case/types';
 import CloseCaseDialog from './CloseCaseDialog';
-import { CaseSectionApi, copyCaseSection } from '../../states/case/sections/api';
-import { lookupApi } from '../../states/case/sections/lookup';
+import { CaseSectionApi} from '../../states/case/sections/api';
+import { lookupApi } from '../../states/case/sections/lookupApi';
+import { copyCaseSection } from '../../states/case/sections/update';
 
 type CaseItemPayload = { [key: string]: string | boolean };
 
@@ -195,7 +196,7 @@ const AddEditCaseItem: React.FC<Props> = ({
       formDefinition.forEach(fd => {
         // A preceding 'filter' call looks nicer but TS type narrowing isn't smart enough to work with that.
         if (fd.type === 'copy-to' && form[fd.name]) {
-          copyCaseSection({
+          newInfo = copyCaseSection({
             definition: definitionVersion,
             original: newInfo,
             fromApi: sectionApi,
