+++ plugin-hrm-form/src/states/case/sections/document.ts
@@ -1,7 +1,9 @@
 import { DefinitionVersion } from 'hrm-form-definitions';
 
-import { CaseSectionApi, getMostRecentSectionItem, getSectionItemById, upsertCaseSectionList } from './api';
+import { CaseSectionApi} from './api';
 import { DocumentEntry } from '../../../types/types';
+import { upsertCaseSectionList } from './update';
+import { getMostRecentSectionItem, getSectionItemById } from './get';
 
 export const documentSectionApi: CaseSectionApi<DocumentEntry> = {
   label: 'Document',
