+++ plugin-hrm-form/src/states/case/sections/household.ts
@@ -1,7 +1,9 @@
 import { DefinitionVersion } from 'hrm-form-definitions';
 
-import { CaseSectionApi, getMostRecentSectionItem, getSectionItemById, upsertCaseSectionList } from './api';
+import { CaseSectionApi} from './api';
 import { HouseholdEntry } from '../../../types/types';
+import { upsertCaseSectionList } from './update';
+import { getMostRecentSectionItem, getSectionItemById } from './get';
 
 export const householdSectionApi: CaseSectionApi<HouseholdEntry> = {
   label: 'Household',
