+++ plugin-hrm-form/src/states/case/sections/perpetrator.ts
@@ -1,7 +1,9 @@
 import { DefinitionVersion } from 'hrm-form-definitions';
 
-import { CaseSectionApi, getMostRecentSectionItem, getSectionItemById, upsertCaseSectionList } from './api';
+import { CaseSectionApi} from './api';
 import { PerpetratorEntry } from '../../../types/types';
+import { upsertCaseSectionList } from './update';
+import { getMostRecentSectionItem, getSectionItemById } from './get';
 
 export const perpetratorSectionApi: CaseSectionApi<PerpetratorEntry> = {
   label: 'Perpetrator',
